Use staticProps and avoid usage of getInitialProps
