+++ next.config.js
@@ -30,5 +30,25 @@ module.exports = withSourceMaps(
     devIndicators: {
       autoPrerender: false,
     },
+    async rewrites() {
+      return [
+        {
+          source: '/',
+          destination: '/de',
+        },
+        {
+          source: '/user/notifications',
+          destination: '/de/user/notifications',
+        },
+        {
+          source: '/user/public',
+          destination: '/de/user/me',
+        },
+        {
+          source: '/user/me',
+          destination: '/de/user/me',
+        },
+      ]
+    },
   })
 )
