+++ src/pages/de/spenden.tsx
@@ -1,0 +1,20 @@
+import { GetStaticProps, NextPage } from 'next'
+import React from 'react'
+
+import { Donations } from '@/components/pages/donations'
+
+export default function Page() {
+  return <Donations />
+}
+
+// eslint-disable-next-line @typescript-eslint/require-await
+export const getStaticProps: GetStaticProps = async () => {
+  return {
+    props: {
+      pageData: {
+        kind: 'donation',
+      },
+      lang: 'de',
+    },
+  }
+}
